Refactor to utils, add test/train/split from .csv
