+++ kneejerk/data/loader.py
@@ -1,5 +1,10 @@
 import cv2
+import click
 import numpy as np
+import pandas as pd
+from sklearn.model_selection import train_test_split
+
+from kneejerk.data.utils import _get_max_image_dim, _determine_pad_amount
 
 
 def load_normalized_image_data(fpath, consider_size=False, rescale_len=200):
@@ -81,43 +86,24 @@ def load_and_pad_images(impath, max_image_dim=None):
     return im
 
 
-def _get_max_image_dim(fpath):
+@click.pass_context
+def segment_data_from_csv(ctx, trainpct, testpct, valpct):
     """
-    Open up all of the images to see their heights and widths.
-    Keep a running max of each, which is returned at the end.
+    Load the csv generated by `kneejerk score` and stratify into
+    test, train, and validation DataFrames, per the parameters
+    provided in `kneejerk transfer`
+
+    :return: train, test, cross_val DataFrames
     """
+    file_name = ctx.obj['file_name']
 
-    max_height = 0
-    max_width = 0
+    df = pd.read_csv(file_name, names=['path', 'score'])
 
-    with open(fpath) as f:
-        for row in f:
-            im_path = row.split(',')[0]
-            im = cv2.imread(im_path)
-            height, width, _ = im.shape
+    train, test = train_test_split(df, train_size=trainpct, stratify=df['score'])
 
-            if height > max_height:
-                max_height = height
+    if valpct:
+        # rescale the original pct
+        testpct = (testpct / (1 - trainpct))
+        test, cross_val = train_test_split(test, train_size=testpct, stratify=test['score'])
 
-            if width > max_width:
-                max_width = width
-
-    print('Max height:', max_height)
-    print('Max width :', max_width)
-
-    return max(max_height, max_width)
-
-
-def _determine_pad_amount(image_array, max_height, max_width):
-    """
-    Given an image and desired max height/width,
-    find how much height and width is needed to make
-    the image the appropriate size
-    """
-
-    height, width, _ = image_array.shape
-
-    height_needed = max_height - height
-    width_needed = max_width - width
-
-    return height_needed, width_needed
+    return train, test, cross_val
